requirements.txt
google-generativeai==0.6.2
python-pptx==0.6.23
Pillow==10.3.0
pytesseract==0.3.10
google-api-core==2.19.0